regorganize packet structure as outlined in https://github.com/ammaraskar/pyCraft/pull/68
