+++ minecraft/networking/packets/clientbound/handshake/__init__.py
@@ -1,0 +1,6 @@
+# Formerly known as state_handshake_clientbound.
+def get_packets(context):
+    packets = {
+
+    }
+    return packets
